feat: addressing comments
